renames GenericPlan -> QueryPlan
