+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/logical/LogicalPlan.java
@@ -2,9 +2,9 @@ package se.liu.ida.hefquin.engine.queryplan.logical;
 
 import java.util.NoSuchElementException;
 
-import se.liu.ida.hefquin.engine.queryplan.info.GenericPlan;
+import se.liu.ida.hefquin.engine.queryplan.base.QueryPlan;
 
-public interface LogicalPlan extends GenericPlan
+public interface LogicalPlan extends QueryPlan
 {
 	/**
 	 * Returns the root operator of this plan.
