+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/logical/impl/LogicalPlanWithNaryRootImpl.java
@@ -6,12 +6,12 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 import se.liu.ida.hefquin.base.query.ExpectedVariables;
-import se.liu.ida.hefquin.engine.queryplan.info.impl.GenericPlanImpl;
+import se.liu.ida.hefquin.engine.queryplan.base.impl.BaseForQueryPlan;
 import se.liu.ida.hefquin.engine.queryplan.logical.LogicalPlan;
 import se.liu.ida.hefquin.engine.queryplan.logical.LogicalPlanWithNaryRoot;
 import se.liu.ida.hefquin.engine.queryplan.logical.NaryLogicalOp;
 
-public class LogicalPlanWithNaryRootImpl extends GenericPlanImpl
+public class LogicalPlanWithNaryRootImpl extends BaseForQueryPlan
                                          implements LogicalPlanWithNaryRoot
 {
 	private final NaryLogicalOp rootOp;
