+++ hefquin-engine/src/main/java/se/liu/ida/hefquin/engine/queryplan/logical/impl/LogicalPlanWithUnaryRootImpl.java
@@ -3,12 +3,12 @@ package se.liu.ida.hefquin.engine.queryplan.logical.impl;
 import java.util.NoSuchElementException;
 
 import se.liu.ida.hefquin.base.query.ExpectedVariables;
-import se.liu.ida.hefquin.engine.queryplan.info.impl.GenericPlanImpl;
+import se.liu.ida.hefquin.engine.queryplan.base.impl.BaseForQueryPlan;
 import se.liu.ida.hefquin.engine.queryplan.logical.LogicalPlan;
 import se.liu.ida.hefquin.engine.queryplan.logical.LogicalPlanWithUnaryRoot;
 import se.liu.ida.hefquin.engine.queryplan.logical.UnaryLogicalOp;
 
-public class LogicalPlanWithUnaryRootImpl extends GenericPlanImpl
+public class LogicalPlanWithUnaryRootImpl extends BaseForQueryPlan
                                           implements LogicalPlanWithUnaryRoot
 {
 	private final UnaryLogicalOp rootOp;
